mini_moka::sync

Trait ConcurrentCacheExt

pub trait ConcurrentCacheExt<K, V> {
    // Required method
    fn sync(&self);
}

Provides extra methods that will be useful for testing.

Required Methods

fn sync(&self)

Performs any pending maintenance operations needed by the cache.